Cleanup references to Travis CI in rake tasks
This makes the real task "release" and makes "travis_release" an alias
for it. It also sets the description of "travis_release" to a message
stating that there is a newer task named "release"

The readme is also updated to reflect the change in task used. Some
additional formatting and cleanup was also applied to the readme.