Rework release preparation tasks
This is a follow-up to https://github.com/voxpupuli/modulesync_config/pull/788

Voxpupuli contributors used to update a module version and `rake
changelog` before opening a PR for preparing a new release.  Some
modules now have a REFERENCE.md file that also needs to be updated (by
running `rake reference`), and forgetting to do so will remain
unnoticed until `rake release` complain that the file is not up-to-date
(so after the release PR was reviewed and merged).

This PR move the `changelog` and `reference` tasks from each module's
Rakefile (under modulesync control), to the voxpupuli-release gem,
rename them to `release:porcelain:changelog` and
`release:porcelain:reference` to leave room for new `changelog` and
`reference` tasks indicating they are new deprecated, and recommand
running the new `release:prepare` task that runs them.

This is indented to help maintainers smoothly update their workflow and
provide a framework where additional tasks can be added if need be.